pyproject.toml
[project]
name = "vaspgo"
version = "0.1.0"
description = "基于AutoGen框架开发的多智能体协作，可实现VASP及其衍生软件的自主计算"
authors = [
    {name = "tangkan",email = "ktang@buaa.edu.cn"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "autogen-agentchat (>=0.7.5,<0.8.0)",
    "autogen-ext[chromadb,mcp,openai] (>=0.7.5,<0.8.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pydantic (>=2.12.4,<3.0.0)",
    "paramiko (>=4.0.0,<5.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "aiohttp (>=3.10.0,<4.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "markitdown (>=0.1.3,<0.2.0)",
    "trafilatura (>=2.0.0,<3.0.0)",
    "pymatgen (>=2025.10.7)"
]

[tool.poetry]
packages = [{include = "vaspgo", from = "src"}]

[[tool.poetry.source]]
name = "pypi-tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[[tool.uv.index]]
url="https://pypi.tuna.tsinghua.edu.cn/simple"
default=true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
